requirements.txt
alembic==1.8.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
PyJWT==1.4.2
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.41
uuid==1.30
Werkzeug==2.2.2
zipp==3.8.1
